@@ -1,10 +1,11 @@
 package com.polidea.flutterblelib.wrapper;
 
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.polidea.rxandroidble.RxBleConnection;
-import com.polidea.rxandroidble.RxBleDevice;
+import com.polidea.rxandroidble2.RxBleConnection;
+import com.polidea.rxandroidble2.RxBleDevice;
 
 import java.util.List;
 import java.util.UUID;
