@@ -2,12 +2,13 @@ package com.polidea.flutterblelib.wrapper;
 
 import android.os.ParcelUuid;
 
-import com.polidea.rxandroidble.scan.ScanFilter;
-import com.polidea.rxandroidble.scan.ScanSettings;
+import com.polidea.rxandroidble2.scan.ScanFilter;
+import com.polidea.rxandroidble2.scan.ScanSettings;
+
 
 public class ScanSettingsWrapper {
     private final ScanSettings scanSettings;
-    private final  String[] uuids;
+    private final String[] uuids;
 
     public ScanSettingsWrapper(ScanSettings scanSettings, String[] uuids) {
         this.scanSettings = scanSettings;
@@ -22,7 +23,7 @@ public class ScanSettingsWrapper {
         if (uuids == null) {
             return null;
         }
-        final ScanFilter [] scanFilters = new ScanFilter[uuids.length];
+        final ScanFilter[] scanFilters = new ScanFilter[uuids.length];
         for (int index = 0; index < uuids.length; index++) {
             scanFilters[index] = new ScanFilter.Builder().setServiceUuid(ParcelUuid.fromString(uuids[index])).build();
         }
